# book-network/src/main/java/com/sarra/book/auth/AuthenticationRequest.java
package com.sarra.book.auth;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.Size;
import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationRequest {
    @Email(message = "Email is not formated")
    @NotEmpty(message="Email is mandatory")
    @NotBlank(message="Email is mandatory")
    private String email;

    @Size(min= 8, message = "password should be 8 caracter minimum")
    @NotEmpty(message="Password is mandatory")
    @NotBlank(message="Password is mandatory")
    private String password;
}